Extract Android's KtorHttpServer and align with JVM version.
