+++ gradle/libs.versions.toml
@@ -1,6 +1,4 @@
 [versions]
-appcompatV7 = "28.0.0"
-coreKtx = "1.13.1"
 kotlin = "2.0.21"
 coroutines = "1.9.0"
 serializationRuntime = "1.6.3"
@@ -13,11 +11,11 @@ lwjgl = "3.3.1"
 
 # JVM:
 ktor = "2.3.12"
-
 tarsosDsp = "2.5"
 
 # Android
 androidGradlePlugin = "8.5.2"
+coreKtx = "1.13.1"
 android-compileSdk = "34"
 android-minSdk = "28"
 android-targetSdk = "34"
